Refactor: 템플릿 기반 InputManager 구조 도입 및 키보드 입력 컴포넌트 분리
- InputManager를 템플릿 클래스 형태로 일반화
- KeyboardDevice 특수화를 통해 키보드 입력 처리 기능 분리
- 입력 관련 상수(KEY_PRESSED_FLAG, MAX_KEY_COUNT)를 InputConstants로 분리
- 키보드 입력 구현 파일을 KeyboardInputManager.{h/cpp}로 분할
- 오타 수정 및 네임스페이스 구조 정리

향후 MouseDevice, GamepadDevice 등 다양한 입력 디바이스 확장 예정
