+++ Butterfly/Event/InputConstants.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace Input
+{
+    constexpr int MAX_KEY_COUNT = 256;
+    constexpr SHORT KEY_PRESSED_FLAG = 0x8000;
+}
