+++ Butterfly/Event/KeyboardInputManager.cpp
@@ -1,0 +1,35 @@
+#include "KeyboardInputManager.h"
+#include <cstring>
+
+namespace Input
+{
+    InputManager<KeyboardDevice>& InputManager<KeyboardDevice>::GetInstance()
+    {
+        static InputManager instance;
+        return instance;
+    }
+
+    void InputManager<KeyboardDevice>::Update()
+    {
+        std::memcpy(prevState.data(), currState.data(), sizeof(SHORT) * MAX_KEY_COUNT);
+        for (int i = 0; i < MAX_KEY_COUNT; ++i)
+        {
+            currState[i] = GetAsyncKeyState(i);
+        }
+    }
+
+    bool InputManager<KeyboardDevice>::IsKeyDown(int vKey) const
+    {
+        return (currState[vKey] & KEY_PRESSED_FLAG) != 0;
+    }
+
+    bool InputManager<KeyboardDevice>::IsKeyPressed(int vKey) const
+    {
+        return (!(prevState[vKey] & KEY_PRESSED_FLAG) && (currState[vKey] & KEY_PRESSED_FLAG));
+    }
+
+    bool InputManager<KeyboardDevice>::IsKeyReleased(int vKey) const
+    {
+        return ((prevState[vKey] & KEY_PRESSED_FLAG) && !(currState[vKey] & KEY_PRESSED_FLAG));
+    }
+}
